requirements.txt
pymavlink
ffmpeg-python